changing pw, security issues, video player frontend, start working on recording video w/ personal device

+++ controllers/videoController.js
@@ -39,8 +39,11 @@ export const postUpload = async(req, res) => {
     const newVideo = await Video.create({
         fileUrl: path,
         title,
-        description
+        description,
+        creator: req.user
     });
+    res.locals.loggedUser.videos.push(newVideo.id);
+    res.locals.loggedUser.save();
     res.redirect(routes.videoDetail(newVideo.id));
 };
 
@@ -49,7 +52,7 @@ export const videoDetail = async (req,res) => {
         params: { id }
     } = req;
     try{
-        const video = await Video.findById(id);
+        const video = await Video.findById(id).populate("creator");
         res.render("videoDetail", {pageTitle: video.title, video /* same as doing "video: video" */ }); 
     } catch(error){
         res.redirect(routes.home);
@@ -62,7 +65,11 @@ export const getEditVideo = async (req,res) => {
     }=req;
     try{
         const video = await Video.findById(id);
-        res.render("editVideo", { pageTitle: `Edit ${video.title}`, video});
+        if(video.creator != req.user){
+            throw Error();
+        } else {
+            res.render("editVideo", { pageTitle: `Edit ${video.title}`, video});
+        }
     } catch(error) {
         res.redirect(routes.home);
     }
@@ -86,7 +93,12 @@ export const deleteVideo = async (req,res) => {
         params: { id }
     }=req;
     try{
-        await Video.findOneAndRemove({_id: id});
+        const video = await Video.findById(id);
+        if(video.creator != req.user){
+            throw Error();
+        } else {
+            await Video.findOneAndRemove({_id: id});
+        }
     } catch(error) { 
         console.log(error);
     }
